Update in progress.
* Ran code inspection and fixed the following types of errors:
    - Unused XML resource (such as "tools")
    - Fixed Javadoc errors.
    - Removed unused imports.
    - Fixed various other code warnings.

+++ app/src/main/java/com/nfcalarmclock/NacAutoDismissPreference.java
@@ -72,7 +72,7 @@ public class NacAutoDismissPreference
 	protected Object onGetDefaultValue(TypedArray a, int index)
 	{
 		NacSharedDefaults defaults = new NacSharedDefaults(getContext());
-		return (Integer) a.getInteger(index, defaults.getAutoDismissIndex());
+		return a.getInteger(index, defaults.getAutoDismissIndex());
 	}
 
 	/**
